Controller only restored #1: pior-grosicki remarks included

+++ src/main/java/com/kodilla/ecommercee/service/UserDbService.java
@@ -10,10 +10,8 @@ import com.kodilla.ecommercee.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.server.ResponseStatusException;
 
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
@@ -57,13 +55,19 @@ public class UserDbService {
 
     public void deleteUser(Long id) {
 
-            userRepository.deleteById(id);
+        if (!userRepository.findById(id).isPresent()) {
+            throw new ResponseStatusException(HttpStatus.NOT_ACCEPTABLE, "No such user!");
+        }
+        userRepository.deleteById(id);
     }
-
 
     public User update(User u) {
 
         Long id = u.getId();
+        if (!userRepository.findById(id).isPresent()) {
+            throw new ResponseStatusException(HttpStatus.NOT_ACCEPTABLE, "No such user!");
+        }
+
         User user = userRepository.findById(u.getId()).get();
         Key userKey = keyRepository.findByUserId(id);
